WIP Move landing page code to landing page
